+++ src/org/loklak/api/search/SearchServlet.java
@@ -425,7 +425,7 @@ public class SearchServlet extends HttpServlet {
                         m.setLink(t.getStatusIdUrl().toExternalForm());
                         m.setAuthor(u.getName() + " @" + u.getScreenName());
                         m.setTitle(u.getName() + " @" + u.getScreenName());
-                        m.setDescription(t.getText(shortlink_iflinkexceedslength, shortlink_urlstub).text);
+                        m.setDescription(t.getText(shortlink_iflinkexceedslength, shortlink_urlstub, t.getText(), t.getLinks(), t.getPostId()).text);
                         m.setPubDate(t.getCreatedAt());
                         m.setGuid(t.getPostId());
                         feed.addMessage(m);
@@ -445,7 +445,11 @@ public class SearchServlet extends HttpServlet {
                 try {
                     for (TwitterTweet t : tl) {
                         UserEntry u = tl.getUser(t);
-                        buffer.append(t.getCreatedAt()).append(" ").append(u.getScreenName()).append(": ").append(t.getText(shortlink_iflinkexceedslength, shortlink_urlstub).text).append('\n');
+                        buffer.append(t.getCreatedAt()).append(" ")
+                                .append(u.getScreenName())
+                                .append(": ")
+                                .append(t.getText(shortlink_iflinkexceedslength, shortlink_urlstub, t.getText(), t.getLinks(), t.getPostId()).text)
+                                .append('\n');
                     }
                 } catch (ConcurrentModificationException e) {
                     // late incoming messages from concurrent peer retrieval may cause this
