+++ src/org/loklak/objects/MessageEntry.java
@@ -24,7 +24,7 @@ import java.net.URL;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Date;
-import java.util.LinkedHashSet;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
@@ -48,396 +48,19 @@ import org.unbescape.html.HtmlEscape;
 public class MessageEntry extends AbstractObjectEntry {
 
     public static final String RICH_TEXT_SEPARATOR = "\n***\n";
+    public Map<Context, Classification<String, Category>> classifier;
 
-    protected Date timestampDate;  // a time stamp that is given in loklak upon the arrival of the tweet which is the current local time
-    protected Date created_at; // the time given in the tweet which is the time when the user created it. This is also use to do the index partition into minute, hour, week
-    protected Date on;         // on means 'valid from'
-    protected Date to;         // 'to' means 'valid_until' and may not be set
-    protected SourceType source_type; // where did the message come from
-    protected ProviderType provider_type;  // who created the message
-    protected String provider_hash, screen_name, retweet_from, postId, canonical_id, parent, text;
-    protected URL status_id_url;
-    protected long retweet_count, favourites_count;
-    protected LinkedHashSet<String> images, audio, videos;
-    protected String place_name, place_id;
-
-    // the following fields are either set as a common field or generated by extraction from field 'text' or from field 'place_name'
-    protected double[] location_point, location_mark; // coordinate order is [longitude, latitude]
-    protected int location_radius; // meter
-    protected LocationSource location_source;
-    protected PlaceContext place_context;
-    protected String place_country;
-    private boolean enriched;
-
-    // the following can be computed from the tweet data but is stored in the search index to provide statistical data and ranking attributes
-    private int without_l_len, without_lu_len, without_luh_len; // the length of tweets without links, users, hashtags
-    private String[] hosts, links, mentions, hashtags; // the arrays of links, users, hashtags
-    private Map<Context, Classification<String, Category>> classifier;
-
-    public MessageEntry() throws MalformedURLException {
-        this.timestamp = new Date().getTime();
-        this.timestampDate = new Date(this.timestamp);
-        this.created_at = new Date();
-        this.on = null;
-        this.to = null;
-        this.source_type = SourceType.GENERIC;
-        this.provider_type = ProviderType.NOONE;
-        this.provider_hash = "";
-        this.screen_name = "";
-        this.retweet_from = "";
-        this.postId = "";
-        this.canonical_id = "";
-        this.parent = "";
-        this.text = "";
-        this.status_id_url = null;
-        this.retweet_count = 0;
-        this.favourites_count = 0;
-        this.images = new LinkedHashSet<String>();
-        this.audio = new LinkedHashSet<String>();
-        this.videos = new LinkedHashSet<String>();
-        this.place_id = "";
-        this.place_name = "";
-        this.place_context = null;
-        this.place_country = null;
-        this.location_point = null;
-        this.location_radius = 0;
-        this.location_mark = null;
-        this.location_source = null;
-        this.without_l_len = 0;
-        this.without_lu_len = 0;
-        this.without_luh_len = 0;
-        this.hosts = new String[0];
-        this.links = new String[0];
-        this.mentions = new String[0];
-        this.hashtags = new String[0];
-        this.classifier = null;
-        this.enriched = false;
-    }
-
-    public MessageEntry(JSONObject json) {
-        Object timestamp_obj = lazyGet(json, AbstractObjectEntry.TIMESTAMP_FIELDNAME);
-        this.timestampDate = parseDate(timestamp_obj);
-        this.timestamp = this.timestampDate.getTime();
-        Object created_at_obj = lazyGet(json, AbstractObjectEntry.CREATED_AT_FIELDNAME);
-        this.created_at = parseDate(created_at_obj);
-        Object on_obj = lazyGet(json, "on");
-        this.on = on_obj == null ? null : parseDate(on);
-        Object to_obj = lazyGet(json, "to");
-        this.to = to_obj == null ? null : parseDate(to);
-        String source_type_string = (String) lazyGet(json, "source_type");
-        try {
-            this.source_type = source_type_string == null ? SourceType.GENERIC : SourceType.byName(source_type_string);
-        } catch (IllegalArgumentException e) {
-            this.source_type = SourceType.GENERIC;
-        }
-        String provider_type_string = (String) lazyGet(json, "provider_type");
-        if (provider_type_string == null) provider_type_string = ProviderType.NOONE.name();
-        try {
-            this.provider_type = ProviderType.valueOf(provider_type_string);
-        } catch (IllegalArgumentException e) {
-            this.provider_type = ProviderType.NOONE;
-        }
-        this.provider_hash = (String) lazyGet(json, "provider_hash");
-        this.screen_name = (String) lazyGet(json, "screen_name");
-        this.retweet_from = (String) lazyGet(json, "retweet_from");
-        this.postId = (String) lazyGet(json, "post_id");
-        this.text = (String) lazyGet(json, "text");
-        try {
-            this.status_id_url = new URL((String) lazyGet(json, "link"));
-        } catch (MalformedURLException e) {
-            this.status_id_url = null;
-        }
-        this.retweet_count = parseLong((Number) lazyGet(json, "retweet_count"));
-        this.favourites_count = parseLong((Number) lazyGet(json, "favourites_count"));
-        this.images = parseArrayList(lazyGet(json, "images"));
-        this.audio = parseArrayList(lazyGet(json, "audio"));
-        this.videos = parseArrayList(lazyGet(json, "videos"));
-        this.place_id = parseString((String) lazyGet(json, "place_id"));
-        this.place_name = parseString((String) lazyGet(json, "place_name"));
-        this.place_country = parseString((String) lazyGet(json, "place_country"));
-        if (this.place_country != null && this.place_country.length() != 2) this.place_country = null;
-
-        // optional location
-        Object location_point_obj = lazyGet(json, "location_point");
-        Object location_radius_obj = lazyGet(json, "location_radius");
-        Object location_mark_obj = lazyGet(json, "location_mark");
-        Object location_source_obj = lazyGet(json, "location_source");
-        if (location_point_obj == null || location_mark_obj == null ||
-            !(location_point_obj instanceof List<?>) ||
-            !(location_mark_obj instanceof List<?>)) {
-            this.location_point = null;
-            this.location_radius = 0;
-            this.location_mark = null;
-            this.location_source = null;
-        } else {
-            this.location_point = new double[]{(Double) ((List<?>) location_point_obj).get(0), (Double) ((List<?>) location_point_obj).get(1)};
-            this.location_radius = (int) parseLong((Number) location_radius_obj);
-            this.location_mark = new double[]{(Double) ((List<?>) location_mark_obj).get(0), (Double) ((List<?>) location_mark_obj).get(1)};
-            this.location_source = LocationSource.valueOf((String) location_source_obj);
-        }
-        this.enriched = false;
-
-        // load enriched data
-        enrich();
-    }
-
-    public Date getTimestampDate() {
-        return this.timestampDate == null ? new Date() : this.timestampDate;
-    }
-
-    public Date getCreatedAt() {
-        return this.created_at == null ? new Date() : this.created_at;
-    }
-
-    public void setCreatedAt(Date created_at) {
-        this.created_at = created_at;
-    }
-
-    public Date getOn() {
-        return this.on;
-    }
-
-    public void setOn(Date on) {
-        this.on = on;
-    }
-
-    public Date getTo() {
-        return this.to;
-    }
-
-    public void setTo(Date to) {
-        this.to = to;
-    }
-
-    public SourceType getSourceType() {
-        return this.source_type;
-    }
-
-    public void setSourceType(SourceType source_type) {
-        this.source_type = source_type;
-    }
-
-    public ProviderType getProviderType() {
-        return provider_type;
-    }
-
-    public void setProviderType(ProviderType provider_type) {
-        this.provider_type = provider_type;
-    }
-
-    public String getProviderHash() {
-        return provider_hash;
-    }
-
-    public void setProviderHash(String provider_hash) {
-        this.provider_hash = provider_hash;
-    }
-
-    public String getScreenName() {
-        return screen_name;
-    }
-
-    public void setScreenName(String user_screen_name) {
-        this.screen_name = user_screen_name;
-    }
-
-    public String getRetweetFrom() {
-        return this.retweet_from;
-    }
-
-    public void setRetweetFrom(String retweet_from) {
-        this.retweet_from = retweet_from;
-    }
-
-    public URL getStatusIdUrl() {
-        return this.status_id_url;
-    }
-
-    public void setStatusIdUrl(URL status_id_url) {
-        this.status_id_url = status_id_url;
-    }
-
-    public long getRetweetCount() {
-        return retweet_count;
-    }
-
-    public void setRetweetCount(long retweet_count) {
-        this.retweet_count = retweet_count;
-    }
-
-    public long getFavouritesCount() {
-        return this.favourites_count;
-    }
-
-    public void setFavouritesCount(long favourites_count) {
-        this.favourites_count = favourites_count;
-    }
-
-    public String getPlaceName() {
-        return place_name;
-    }
-
-    public PlaceContext getPlaceContext () {
-        return place_context;
-    }
-
-    public void setPlaceName(String place_name, PlaceContext place_context) {
-        this.place_name = place_name;
-        this.place_context = place_context;
-    }
-
-    public String getPlaceId() {
-        return place_id;
-    }
-
-    public void setPlaceId(String place_id) {
-        this.place_id = place_id;
-    }
-
-    /**
-     * @return [longitude, latitude]
-     */
-    public double[] getLocationPoint() {
-        return location_point;
-    }
-
-    /**
-     * set the location
-     * @param location_point in the form [longitude, latitude]
-     */
-    public void setLocationPoint(double[] location_point) {
-        this.location_point = location_point;
-    }
-
-    //TODO: to implement this method
-    private void setPostId() {
-        this.postId = String.valueOf(this.timestamp) + String.valueOf(this.created_at.getTime());
-    }
-
-    public String getPostId() {
-        return String.valueOf(this.postId);
-    }
-
-    /**
-     * @return [longitude, latitude] which is inside of getLocationRadius() from getLocationPoint()
-     */
-    public double[] getLocationMark() {
-        return location_mark;
-    }
-
-    /**
-     * set the location
-     * @param location_point in the form [longitude, latitude]
-     */
-    public void setLocationMark(double[] location_mark) {
-        this.location_mark = location_mark;
-    }
-
-    /**
-     * get the radius in meter
-     * @return radius in meter around getLocationPoint() (NOT getLocationMark())
-     */
-    public int getLocationRadius() {
-        return location_radius;
-    }
-
-    public void setLocationRadius(int location_radius) {
-        this.location_radius = location_radius;
-    }
-
-    public LocationSource getLocationSource() {
-        return location_source;
-    }
-
-    public void setLocationSource(LocationSource location_source) {
-        this.location_source = location_source;
-    }
-
-    public void setText(String text) {
-        this.text = text;
-    }
-
-    public void setImages(ArrayList<String> images) {
-        this.images = parseArrayList(images);
-    }
-
-    public void setImages(String[] images) {
-        this.images = parseArrayList(images);
-    }
-
-    public void setImages(String image) {
-        this.images = parseArrayList(image);
-    }
-
-    public long getId() {
-        return Long.parseLong(this.postId);
-    }
-
-    public String[] getHosts() {
-        return this.hosts;
-    }
-
-    public String getText() {
-        return this.text;
-    }
-
-    public TextLinkMap getText(final int iflinkexceedslength, final String urlstub) {
-        // check if we shall replace shortlinks
-        TextLinkMap tlm = new TextLinkMap();
-        tlm.text = this.text;
-        String[] links = this.getLinks();
-        if (links != null) {
-            for (int nth = 0; nth < links.length; nth++) {
-                String link = links[nth];
-                if (link.length() > iflinkexceedslength) {
-                    //if (!DAO.existMessage(this.getPostId())) break linkloop;
-                    String shortlink = urlstub + "/x?id=" + this.getPostId() +
-                        (nth == 0 ? "" : ShortlinkFromTweetServlet.SHORTLINK_COUNTER_SEPERATOR + Integer.toString(nth));
-                    if (shortlink.length() < link.length()) {
-                        tlm.text = tlm.text.replace(link, shortlink);
-                        if (!shortlink.equals(link)) {
-                            int stublen = shortlink.length() + 3;
-                            if (link.length() >= stublen) link = link.substring(0, shortlink.length()) + "...";
-                            tlm.short2long.put(shortlink, link);
-                        }
-                    }
-                }
-            }
-        }
-        return tlm;
-    }
-
-    public static class TextLinkMap {
-        public String text;
-        public JSONObject short2long;
-        public TextLinkMap() {
-            text = "";
-            this.short2long = new JSONObject(true);
-        }
-        public String toString() {
-            return this.text;
-        }
-    }
-
-    public int getTextLength() {
-        return this.text.length();
-    }
-
-    public String[] getMentions() {
-        return this.mentions;
-    }
-
-    public String[] getHashtags() {
-        return this.hashtags;
-    }
-
-    public String[] getLinks() {
-        return this.links;
-    }
-
-    public Collection<String> getImages() {
-        return this.images;
-    }
+    // two or more
+    public final static Pattern SPACEX_PATTERN = Pattern.compile("  +");
+    // right boundary must be space or ) since others may appear in urls
+    public final static Pattern URL_PATTERN = Pattern.compile("(?:\\b|^)(https?://[-A-Za-z0-9+&@#/%?=~_()|!:,.;]*[-A-Za-z0-9+&@#/%=~_()|])");
+    // left boundary must be space since the @ is itself a boundary
+    public final static Pattern USER_PATTERN = Pattern.compile("(?:[ (]|^)(@..*?)(?:\\b|$)"); 
+    // left boundary must be a space since the # is itself a boundary
+    final static Pattern HASHTAG_PATTERN = Pattern.compile("(?:[ (]|^)(#..*?)(?:\\b|$)");
+    public final static Pattern A_END_TAG = Pattern.compile("</a>");
+    public final static Pattern QUOT_TAG = Pattern.compile("&quot;");
+    public final static Pattern AMP_TAG = Pattern.compile("&amp;");
 
     public Classifier.Category getClassifier(Classifier.Context context) {
         if (this.classifier == null) return null;
@@ -453,200 +76,43 @@ public class MessageEntry extends AbstractObjectEntry {
         return classification.getProbability();
     }
 
-    final static Pattern SPACEX_PATTERN = Pattern.compile("  +"); // two or more
-    final static Pattern URL_PATTERN = Pattern.compile("(?:\\b|^)(https?://[-A-Za-z0-9+&@#/%?=~_()|!:,.;]*[-A-Za-z0-9+&@#/%=~_()|])"); // right boundary must be space or ) since others may appear in urls
-    final static Pattern USER_PATTERN = Pattern.compile("(?:[ (]|^)(@..*?)(?:\\b|$)"); // left boundary must be space since the @ is itself a boundary
-    final static Pattern HASHTAG_PATTERN = Pattern.compile("(?:[ (]|^)(#..*?)(?:\\b|$)"); // left boundary must be a space since the # is itself a boundary
+    public static List<String> extract(String text, Pattern p, int regexGroup) {
+        Matcher m = p.matcher(text);
+        List<String> dataList = new ArrayList<String>();
 
-    /**
-     * create enriched data, useful for analytics and ranking:
-     * - identify all mentioned users, hashtags and links
-     * - count message size without links
-     * - count message size without links and without users
-     */
-    public void enrich() {
-        if (this.enriched) return;
-        StringBuilder t = new StringBuilder(this.text);
+        while (m.find()) {
+            dataList.add(m.group(regexGroup));
+        }
+        for (String r: dataList) {
+            text.replaceAll(r, "");
+        }
+        return dataList;
+    }
 
-        // extract the links
-        List<String> links = extract(t, URL_PATTERN, 1);
-        t = new StringBuilder(SPACEX_PATTERN.matcher(t).replaceAll(" ").trim());
-        this.without_l_len = t.length(); // len_no_l
+   public List<String> extractLinks(String text) {
+        return extract(text, URL_PATTERN, 1);
+    }
 
-        // extract the users
-        List<String> users = extract(t, USER_PATTERN, 1);
-        t = new StringBuilder(SPACEX_PATTERN.matcher(t).replaceAll(" ").trim());
-        this.without_lu_len = t.length(); // len_no_l_and_users
+    public List<String> extractUsers(String text) {
+        return extract(text, USER_PATTERN, 1);
+    }
 
-        // extract the hashtags
-        List<String> hashtags = extract(t, HASHTAG_PATTERN, 1);
-        t = new StringBuilder(SPACEX_PATTERN.matcher(t).replaceAll(" ").trim());
-        this.without_luh_len = t.length(); // len_no_l_and_users_and_hashtags
+    public List<String> extractHashtags(String text) {
+        return extract(text, HASHTAG_PATTERN, 1);
+    }
 
-        // extract the hosts from the links
-        Set<String> hosts = new LinkedHashSet<String>();
+    public List<String> extractHosts(List<String> links) {
+        List<String> hosts = new ArrayList<String>();
         for (String u: links) {
             try {
                 URL url = new URL(u);
-                hosts.add(url.getHost());
+                hosts.add(url.getHost().toLowerCase());
             } catch (MalformedURLException e) {}
         }
-
-        this.hosts = new String[hosts.size()];
-        int j = 0; for (String host: hosts) this.hosts[j++] = host.toLowerCase();
-
-        this.mentions = new String[users.size()];
-        for (int i = 0; i < users.size(); i++) this.mentions[i] = users.get(i).substring(1);
-
-        this.hashtags = new String[hashtags.size()];
-        for (int i = 0; i < hashtags.size(); i++) this.hashtags[i] = hashtags.get(i).substring(1).toLowerCase();
-
-        this.links = new String[links.size()];
-        for (int i = 0; i < links.size(); i++) this.links[i] = links.get(i);
-
-        // classify content
-        this.classifier = Classifier.classify(this.text);
-
-        // more media data, analyze the links
-        for (String link: this.links) {
-            if (link.endsWith(".mp4") || link.endsWith(".m4v") ||
-                link.indexOf("vimeo.com") > 0 ||
-                link.indexOf("youtube.com") > 0 || link.indexOf("youtu.be") > 0 ||
-                link.indexOf("vine.co") > 0 ||
-                link.indexOf("ted.com") > 0) {this.videos.add(link); continue;}
-            if (link.endsWith(".mp3") ||
-                link.indexOf("soundcloud.com") > 0) {this.audio.add(link); continue;}
-            if (link.endsWith(".jpg") ||
-                link.endsWith(".jpeg") ||
-                link.endsWith(".png") ||
-                link.endsWith(".gif") ||
-                link.indexOf("flickr.com") > 0 ||
-                link.indexOf("instagram.com") > 0 ||
-                link.indexOf("imgur.com") > 0 ||
-                link.indexOf("giphy.com") > 0 ||
-                link.indexOf("pic.twitter.com") > 0) {this.images.add(link); continue;}
-        }
-
-        // find location
-        if ((this.location_point == null || this.location_point.length == 0) && DAO.geoNames != null) {
-            GeoMark loc = null;
-            if (this.place_name != null && this.place_name.length() > 0 &&
-                (this.location_source == null || this.location_source == LocationSource.ANNOTATION || this.location_source == LocationSource.PLACE)) {
-                loc = DAO.geoNames.analyse(this.place_name, null, 5, Integer.toString(this.text.hashCode()));
-                this.place_context = PlaceContext.FROM;
-                this.location_source = LocationSource.PLACE;
-            }
-            if (loc == null) {
-                loc = DAO.geoNames.analyse(this.text, this.hashtags, 5, Integer.toString(this.text.hashCode()));
-                this.place_context = PlaceContext.ABOUT;
-                this.location_source = LocationSource.ANNOTATION;
-            }
-            if (loc != null) {
-                if (this.place_name == null || this.place_name.length() == 0) this.place_name = loc.getNames().iterator().next();
-                this.location_radius = 0;
-                this.location_point = new double[]{loc.lon(), loc.lat()}; //[longitude, latitude]
-                this.location_mark = new double[]{loc.mlon(), loc.mlat()}; //[longitude, latitude]
-                this.place_country = loc.getISO3166cc();
-            }
-        }
-        this.enriched = true;
+        return hosts;
     }
 
-    private static List<String> extract(StringBuilder s, Pattern p, int g) {
-        Matcher m = p.matcher(s.toString());
-        List<String> l = new ArrayList<String>();
-        while (m.find()) l.add(m.group(g));
-        for (String r: l) {int i = s.indexOf(r); s.replace(i, i + r.length(), "");}
-        return l;
-    }
-
-    @Override
     public JSONObject toJSON() {
-        return toJSON(null, true, Integer.MAX_VALUE, ""); // very important to include calculated data here because that is written into the index using the abstract index factory
-    }
-
-    public Post toJSON(final UserEntry user, final boolean calculatedData, final int iflinkexceedslength, final String urlstub) {
-
-        // tweet data
-        this.put(AbstractObjectEntry.TIMESTAMP_FIELDNAME, utcFormatter.print(getTimestampDate().getTime()));
-        this.put(AbstractObjectEntry.CREATED_AT_FIELDNAME, utcFormatter.print(getCreatedAt().getTime()));
-        if (this.on != null) this.put("on", utcFormatter.print(this.on.getTime()));
-        if (this.to != null) this.put("to", utcFormatter.print(this.to.getTime()));
-        this.put("screen_name", this.screen_name);
-        if (this.retweet_from != null && this.retweet_from.length() > 0) this.put("retweet_from", this.retweet_from);
-        TextLinkMap tlm = this.getText(iflinkexceedslength, urlstub);
-        this.put("text", tlm); // the tweet; the cleanup is a helper function which cleans mistakes from the past in scraping
-        if (this.status_id_url != null) this.put("link", this.status_id_url.toExternalForm());
-        this.put("id_str", this.postId);
-        if (this.canonical_id != null) this.put("canonical_id", this.canonical_id);
-        if (this.parent != null) this.put("parent", this.parent);
-        this.put("source_type", this.source_type.toString());
-        this.put("provider_type", this.provider_type.name());
-        if (this.provider_hash != null && this.provider_hash.length() > 0) this.put("provider_hash", this.provider_hash);
-        this.put("retweet_count", this.retweet_count);
-        this.put("favourites_count", this.favourites_count); // there is a slight inconsistency here in the plural naming but thats how it is noted in the twitter api
-        this.put("place_name", this.place_name);
-        this.put("place_id", this.place_id);
-
-        // add statistic/calculated data
-        if (calculatedData) {
-
-            // text length
-            this.put("text_length", this.text.length());
-
-            // location data
-            if (this.place_context != null) this.put("place_context", this.place_context.name());
-            if (this.place_country != null && this.place_country.length() == 2) {
-                this.put("place_country", DAO.geoNames.getCountryName(this.place_country));
-                this.put("place_country_code", this.place_country);
-                this.put("place_country_center", DAO.geoNames.getCountryCenter(this.place_country));
-            }
-
-            // add optional location data. This is written even if calculatedData == false if the source is from REPORT to prevent that it is lost
-            if (this.location_point != null && this.location_point.length == 2 && this.location_mark != null && this.location_mark.length == 2) {
-                // reference for this format: https://www.elastic.co/guide/en/elasticsearch/reference/current/mapping-geo-point-type.html#_lat_lon_as_array_5
-                this.put("location_point", this.location_point); // [longitude, latitude]
-                this.put("location_radius", this.location_radius);
-                this.put("location_mark", this.location_mark);
-                this.put("location_source", this.location_source.name());
-            }
-
-            // redundant data for enhanced navigation with aggregations
-            this.put("hosts", this.hosts);
-            this.put("hosts_count", this.hosts.length);
-            this.put("links", this.links);
-            this.put("links_count", this.links.length);
-            this.put("unshorten", tlm.short2long);
-            this.put("images", this.images);
-            this.put("images_count", this.images.size());
-            this.put("audio", this.audio);
-            this.put("audio_count", this.audio.size());
-            this.put("videos", this.videos);
-            this.put("videos_count", this.videos.size());
-            this.put("mentions", this.mentions);
-            this.put("mentions_count", this.mentions.length);
-            this.put("hashtags", this.hashtags);
-            this.put("hashtags_count", this.hashtags.length);
-
-            // text classifier
-            if (this.classifier != null) {
-                for (Map.Entry<Context, Classification<String, Category>> c: this.classifier.entrySet()) {
-                    assert c.getValue() != null;
-                    if (c.getValue().getCategory() == Classifier.Category.NONE) continue; // we don't store non-existing classifications
-                    this.put("classifier_" + c.getKey().name(), c.getValue().getCategory());
-                    this.put("classifier_" + c.getKey().name() + "_probability",
-                        c.getValue().getProbability() == Float.POSITIVE_INFINITY ? Float.MAX_VALUE : c.getValue().getProbability());
-                }
-            }
-
-            // experimental, for ranking
-            this.put("without_l_len", this.without_l_len);
-            this.put("without_lu_len", this.without_lu_len);
-            this.put("without_luh_len", this.without_luh_len);
-        }
-
-        // add user
-        if (user != null) this.put("user", user.toJSON());
         return this;
     }
 
@@ -698,15 +164,99 @@ public class MessageEntry extends AbstractObjectEntry {
         return clean.toString().trim().replaceAll(" +", " ");
     }
 
-    public Collection<String> getVideos() {
-        return this.videos;
+    public Set<String> getLinksVideo(List<String> links) {
+        Set<String> videoList = new HashSet<String>();
+        boolean endLink = false;
+        boolean withLinkTerm = false;
+
+        for (String link: links) {
+            endLink = link.endsWith(".mp4") || link.endsWith(".m4v");
+            withLinkTerm = link.indexOf("vimeo.com") > 0
+                    || link.indexOf("youtube.com") > 0
+                    || link.indexOf("youtu.be") > 0
+                    || link.indexOf("vine.co") > 0
+                    || link.indexOf("ted.com") > 0;
+
+            if (endLink || withLinkTerm) {
+                videoList.add(link);
+            }
+        }
+        return videoList;
     }
 
-    public Collection<String> getAudio() {
-        return this.audio;
+    public Set<String> getLinksAudio(List<String> links) {
+        Set<String> audioLinks = new HashSet<String>();
+        boolean endLink = false;
+        boolean withLinkTerm = false;
+
+        for (String link: links) {
+            endLink = link.endsWith(".mp3");
+            withLinkTerm = link.indexOf("soundcloud.com") > 0;
+
+            if (withLinkTerm || endLink) audioLinks.add(link);
+        }
+        return audioLinks;
     }
 
-    private final static Pattern A_END_TAG = Pattern.compile("</a>");
-    private final static Pattern QUOT_TAG = Pattern.compile("&quot;");
-    private final static Pattern AMP_TAG = Pattern.compile("&amp;");
+    public Set<String> getLinksImage(List<String> links ) {
+        Set<String> imageLinks = new HashSet<String>();
+        boolean endLink = false;
+        boolean withLinkTerm = false;
+        
+        for (String link: links) {
+            endLink = link.endsWith(".jpg")
+                    || link.endsWith(".jpeg")
+                    || link.endsWith(".png")
+                    || link.endsWith(".gif");
+            /*
+             * TODO: Commented this as not all links are of image. Need to fix this.
+            withLinkTerm = link.indexOf("flickr.com") > 0
+                    || link.indexOf("instagram.com") > 0
+                    || link.indexOf("imgur.com") > 0
+                    || link.indexOf("giphy.com") > 0
+                    || link.indexOf("pic.twitter.com") > 0;
+            */
+            if (endLink || withLinkTerm) imageLinks.add(link);
+        }
+        return imageLinks;
+    }
+
+    public TextLinkMap getText(final int iflinkexceedslength, final String urlstub, String text, String[] linksArray, String postId) {
+        // check if we shall replace shortlinks
+        TextLinkMap tlm = new TextLinkMap();
+        tlm.text = text;
+        String[] links = linksArray;
+        if (links != null) {
+            for (int nth = 0; nth < links.length; nth++) {
+                String link = links[nth];
+                if (link.length() > iflinkexceedslength) {
+                    // Here is bug/ quickfix to solve bug. Need to fix for issue https://github.com/loklak/loklak_server/issues/1346
+                    //if (!DAO.existMessage(this.getPostId())) break;
+                    String shortlink = urlstub + "/x?id=" + postId +
+                        (nth == 0 ? "" : ShortlinkFromTweetServlet.SHORTLINK_COUNTER_SEPERATOR + Integer.toString(nth));
+                    if (shortlink.length() < link.length()) {
+                        tlm.text = tlm.text.replace(link, shortlink);
+                        if (!shortlink.equals(link)) {
+                            int stublen = shortlink.length() + 3;
+                            if (link.length() >= stublen) link = link.substring(0, shortlink.length()) + "...";
+                            tlm.short2long.put(shortlink, link);
+                        }
+                    }
+                }
+            }
+        }
+        return tlm;
+    }
+
+    public static class TextLinkMap {
+        public String text;
+        public JSONObject short2long;
+        public TextLinkMap() {
+            text = "";
+            this.short2long = new JSONObject(true);
+        }
+        public String toString() {
+            return this.text;
+        }
+    }
 }
